Modified return annotation for call_async_comprehension

+++ 0x02-python_async_comprehension/2-measure_runtime.py
@@ -23,7 +23,7 @@ async def measure_runtime() -> float:
 
 
 async def call_async_comprehension(
-        times: int) -> Generator[Callable[[], List[int]], None, None]:
+        times: int) -> Generator[Callable[[], List[float]], None, None]:
     """
     Yields async_comprehension @times times
     :param times: number of times async_comprehension is yielded by the
